requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
requests==2.18.2
nose==1.3.7
Flask-OAuthlib==0.9.4


